feat(trace-tree-node): Moving utils away from traceguards file
