+++ static/app/views/performance/newTraceDetails/traceGuards.tsx
@@ -1,4 +1,3 @@
-import type {TraceSplitResults} from './traceApi/types';
 import type {TraceTree} from './traceModels/traceTree';
 import type {BaseNode} from './traceModels/traceTreeNode/baseNode';
 import type {CollapsedNode} from './traceModels/traceTreeNode/collapsedNode';
@@ -78,60 +77,8 @@ export function isTraceErrorNode(node: BaseNode): node is BaseNode<TraceTree.Tra
   return isTraceError(node.value);
 }
 
-export function isTraceNode(
-  node: BaseNode
-): node is BaseNode<TraceSplitResults<TraceTree.Transaction>> {
+export function isTraceNode(node: BaseNode): boolean {
   return !!(node.value && 'orphan_errors' in node.value && 'transactions' in node.value);
-}
-
-export function shouldAddMissingInstrumentationSpan(sdk: string | undefined): boolean {
-  if (!sdk) {
-    return true;
-  }
-  if (sdk.length < 'sentry.javascript.'.length) {
-    return true;
-  }
-
-  switch (sdk.toLowerCase()) {
-    case 'sentry.javascript.browser':
-    case 'sentry.javascript.react':
-    case 'sentry.javascript.gatsby':
-    case 'sentry.javascript.ember':
-    case 'sentry.javascript.vue':
-    case 'sentry.javascript.angular':
-    case 'sentry.javascript.angular-ivy':
-    case 'sentry.javascript.nextjs':
-    case 'sentry.javascript.nuxt':
-    case 'sentry.javascript.electron':
-    case 'sentry.javascript.remix':
-    case 'sentry.javascript.svelte':
-    case 'sentry.javascript.sveltekit':
-    case 'sentry.javascript.react-native':
-    case 'sentry.javascript.astro':
-      return false;
-    case undefined:
-      return true;
-    default:
-      return true;
-  }
-}
-
-export function isJavascriptSDKEvent(value: TraceTree.NodeValue): boolean {
-  return (
-    !!value &&
-    'sdk_name' in value &&
-    /javascript|angular|astro|backbone|ember|gatsby|nextjs|react|remix|svelte|vue/.test(
-      value.sdk_name
-    )
-  );
-}
-
-export function isBrowserRequestNode(node: BaseNode): boolean {
-  return (
-    // Adjust for SDK changes in https://github.com/getsentry/sentry-javascript/pull/13527
-    node.op === 'browser.request' ||
-    (node.op === 'browser' && node.description === 'request')
-  );
 }
 
 export function isTraceOccurence(
@@ -146,28 +93,4 @@ export function isEAPTraceOccurrence(
   return (
     isTraceOccurence(issue) && 'event_type' in issue && issue.event_type === 'occurrence'
   );
-}
-
-export function isStandaloneSpanMeasurementNode(node: BaseNode): boolean {
-  if (node.value && 'op' in node.value && node.value.op) {
-    if (
-      node.value.op.startsWith('ui.webvital.') ||
-      node.value.op.startsWith('ui.interaction.')
-    ) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-export function isRootEvent(value: TraceTree.NodeValue): boolean {
-  // Root events has no parent_span_id
-  return !!value && 'parent_span_id' in value && value.parent_span_id === null;
-}
-
-export function isTraceSplitResult(
-  result: TraceTree.Trace
-): result is TraceSplitResults<TraceTree.Transaction> {
-  return 'transactions' in result && 'orphan_errors' in result;
 }
